Using GetDetails instead of Search, error catching
The search function doesn't return much information about the button and
a large amount had to be hard coded. In comes GetDetails and its magical
all-retrieving gifts of the gods.
Also added some catching code to make sure the button exists and that
the creation was a success before setting any meta data.